requirements.txt
chromadb==0.4.22
streamlit==1.29.0
pdfplumber==0.10.3
python-docx==1.1.0
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
numpy>=1.24.0
